requirements.txt
django~=5.0.1
psycopg2-binary
psycopg2
pillow
djangorestframework~=3.14.0
getenv
python-dotenv~=1.0.0
utils~=1.0.2
redis
djangorestframework-simplejwt
django_filter
bcrypt
